Move tests to package
