+++ src/test/kotlin/io/github/freya022/botcommands/test/commands/slash/userapps/SlashIntegrationOwners.kt
@@ -1,4 +1,4 @@
-package io.github.freya022.botcommands.test.commands.slash
+package io.github.freya022.botcommands.test.commands.slash.userapps
 
 import dev.minn.jda.ktx.coroutines.await
 import dev.minn.jda.ktx.messages.into
